requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.2
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
coverage==5.4
cryptography==3.4.5
csv-diff==1.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dictdiffer==0.8.1
distlib==0.3.1
Django==3.0.5
django-filter==2.4.0
django-widget-tweaks==1.4.8
docopt==0.6.2
dominate==2.6.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geopy==2.1.0
gurobipy==9.1.1
html5lib==1.1
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
iniconfig==1.1.1
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lesscpy==0.14.0
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
mysql-connector==2.2.9
mysql-connector-python==8.0.23
mysqlclient==2.0.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.5
newsapi-python==0.2.6
nltk==3.5
notebook==6.2.0
numpy==1.20.1
packaging==20.8
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.14.0
py==1.10.0
pycoingecko==2.0.0
pycparser==2.20
Pygments==2.8.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.2
pymongo==3.11.3
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
pytest==6.2.2
python-bidi==0.4.2
python-constraint==1.4.0
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
ratelim==0.1.6
regex==2020.11.13
reportlab==3.5.66
requests==2.25.1
rsome==0.1.1
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
sqlparse==0.4.1
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.56.2
traitlets==5.0.5
twilio==6.58.0
urllib3==1.26.4
virtualenv==20.3.1
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
xhtml2pdf==0.2.5
xlrd==2.0.1
zope.interface==5.2.0
